requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
python-decouple==3.8
Pillow==11.0.0
django-filter==24.3
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
